Unify K-Means helpers under tzrec/modules/sid_generation/kmeans.py
Move ``_kmeans`` and ``_residual_kmeans`` (Lloyd's K-Means + residual
variant used by ResidualQuantized.init_embed_) from residual_quantized.py
into kmeans.py, next to ``_kmeans_plus_plus`` / ``_squared_euclidean_distance``
and the ``KMeansLayer`` centroid container. After this move:

  - kmeans.py is the single home for torch-native K-Means code in the
    SID-generation stack — distance helper, KMeans++ seeding, Lloyd's
    iterator, residual variant, and the per-layer centroid container.
  - residual_quantized.py is a quantizer module; it now imports the
    single function it needs (``_residual_kmeans``) instead of pulling
    in two private helpers and re-implementing Lloyd's on top.
  - No functional change. RQ-VAE warm-start remains pure torch / GPU /
    dependency-free — FAISS was not introduced here on purpose: the
    init runs once on ~2k × 64 encoder outputs, where FAISS would add
    a CPU round-trip + a hard dep for no measurable win.

Verified: 24/24 unit tests pass; FAISS smoke test on
forge_contrastive_item_embedding.csv produces bit-identical codes and
RQ-VAE losses.
