+++ tzrec/modules/sid_generation/residual_quantized.py
@@ -11,93 +11,20 @@
 
 """ResidualQuantized: multi-layer residual vector quantization with VQ layers."""
 
-from typing import List, Optional, Tuple, Union
+from typing import List, Optional, Union
 
 import torch
 import torch.distributed as dist
 from torch import nn
 from torch.nn import functional as F
 
-from tzrec.modules.sid_generation.kmeans import (
-    _kmeans_plus_plus,
-    _squared_euclidean_distance,
-)
+from tzrec.modules.sid_generation.kmeans import _residual_kmeans
 from tzrec.modules.sid_generation.types import (
     QuantizeForwardMode,
     QuantizeOutput,
     ResidualQuantizedOutput,
 )
 from tzrec.modules.sid_generation.vector_quantize import VectorQuantize
-
-
-
-
-@torch.no_grad()
-def _kmeans(
-    samples: torch.Tensor,
-    n_clusters: int,
-    n_iters: int = 100,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """Lloyd's K-Means algorithm with KMeans++ initialization.
-
-    Args:
-        samples (Tensor): data points, shape (N, D).
-        n_clusters (int): number of clusters K.
-        n_iters (int): number of iterations. Default: 100.
-
-    Returns:
-        centroids (Tensor): cluster centers, shape (K, D).
-        assignments (Tensor): cluster indices, shape (N,).
-    """
-    N, D = samples.shape
-    centroids = _kmeans_plus_plus(samples, n_clusters)
-
-    for _ in range(n_iters):
-        dists = _squared_euclidean_distance(samples, centroids)  # (N, K)
-        assignments = dists.argmin(dim=-1)  # (N,)
-
-        bins = torch.bincount(assignments, minlength=n_clusters)
-        zero_mask = bins == 0
-        bins_clamped = bins.masked_fill(zero_mask, 1)
-
-        new_centroids = torch.zeros_like(centroids)
-        new_centroids.scatter_add_(
-            0, assignments.unsqueeze(1).expand(-1, D), samples
-        )
-        new_centroids = new_centroids / bins_clamped.unsqueeze(1)
-
-        # Keep old centroids for empty clusters
-        centroids = torch.where(
-            zero_mask.unsqueeze(1), centroids, new_centroids
-        )
-
-    return centroids, assignments
-
-
-@torch.no_grad()
-def _residual_kmeans(
-    samples: torch.Tensor,
-    n_clusters_list: List[int],
-    n_iters: int = 100,
-) -> List[torch.Tensor]:
-    """Residual K-Means: sequentially cluster and subtract centroids.
-
-    Args:
-        samples (Tensor): data points, shape (N, D).
-        n_clusters_list (List[int]): per-layer cluster counts.
-        n_iters (int): K-Means iterations. Default: 100.
-
-    Returns:
-        List[Tensor]: per-layer centroids [(K0, D), (K1, D), ...].
-    """
-    res_centers = []
-    for n_clusters in n_clusters_list:
-        centroids, assignments = _kmeans(samples, n_clusters, n_iters)
-        res_centers.append(centroids)
-        samples = samples - centroids[assignments]
-    return res_centers
-
-
 
 
 class ResidualQuantized(nn.Module):
